pyproject.toml
[project]
name = "projectglamira"
version = "0.1.0"
description = ""
authors = [
    {name = "huytq1412",email = "huytrinhquang1412@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "faker (>=40.1.0,<41.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "ip2location (>=8.11.0,<9.0.0)",
    "html5lib (>=1.1,<2.0)",
    "curl-cffi (>=0.14.0,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
